Cleanup: moved auth middleware to new middleware folder
